feat(add insertion sort): Insertion sort is a step more complex but a bit more useful than bubble so
Insertion sort is a step more complex but a bit more useful than bubble sort and is occasionally

useful. The worst case scenario for it is similar to bubble sort's but its best case makes it suited

for times when you're pretty sure a list almost sorted or likely already sorted.

What's the Big O? There's an inner loop that goes over your sorted list to find the correct place to

insert your item, and an outer loop to go over all the numbers. Two loops means O(n²). However since

if your list is sorted or nearly so, it can be O(n) in a best case scenario and thus well adapted to

that scenario.